Broke up Scte35 module into Scte35 package
