+++ scte/Scte35/SpliceInsert.py
@@ -1,0 +1,3 @@
+class SpliceInsert:
+    def __init__(self, bitarray_data):
+        None
