+++ scte/Scte35/SpliceNull.py
@@ -1,0 +1,3 @@
+class SpliceNull:
+    def __init__(self, bitarray_data):
+        None
