+++ scte/Scte35/SpliceSchedule.py
@@ -1,0 +1,3 @@
+class SpliceSchedule:
+    def __init__(self, bitarray_data):
+        None
