+++ scte/Scte35/__init__.py
@@ -1,0 +1,7 @@
+from scte.Scte35.SegmentationDescriptor import SegmentationDescriptor
+from scte.Scte35.SpliceDescriptor import SpliceDescriptor
+from scte.Scte35.SpliceEvent import SpliceEvent
+from scte.Scte35.SpliceInsert import SpliceInsert
+from scte.Scte35.SpliceNull import SpliceNull
+from scte.Scte35.SpliceSchedule import SpliceSchedule
+from scte.Scte35.TimeSignal import TimeSignal
