+++ scte/__init__.py
@@ -1,2 +1,1 @@
-from scte.Scte35 import Scte35
-from scte import scte35_enums
+from scte import Scte35
